requirements.txt
requests==2.21.0
virtualenv==16.3.0
python-dotenv==0.15.0